refactor(shared): extract verification email queue and auth event types
