+++ apps/api/src/auth/auth-events.ts
@@ -1,28 +1,8 @@
-import { appendFile, mkdir } from "node:fs/promises";
-import path from "node:path";
-
-export type AuthDomainEventName = "user.registered" | "user.verification_sent";
-
-export type AuthDomainEvent = {
-  eventId: string;
-  name: AuthDomainEventName;
-  occurredAt: string;
-  actor: {
-    userId: string;
-    workspaceId: null;
-  };
-  data: Record<string, unknown>;
-};
-
-const defaultEventsFilePath = () => path.resolve(process.cwd(), ".tmp", "domain-events.ndjson");
-
-export const appendAuthDomainEvent = async ({
-  event,
-  eventsFilePath = defaultEventsFilePath(),
-}: {
-  event: AuthDomainEvent;
-  eventsFilePath?: string;
-}) => {
-  await mkdir(path.dirname(eventsFilePath), { recursive: true });
-  await appendFile(eventsFilePath, `${JSON.stringify(event)}\n`, "utf8");
-};
+export {
+  appendAuthDomainEvent,
+  appendAuthDomainEventDeadLetter,
+  defaultAuthEventsDeadLetterFilePath,
+  defaultAuthEventsFilePath,
+  type AuthDomainEvent,
+  type AuthDomainEventName,
+} from "../../../../packages/shared/src/events/auth-events.js";
