+++ apps/api/src/auth/verification-email-queue.ts
@@ -1,151 +1,14 @@
 import { createCipheriv, createHash, randomBytes } from "node:crypto";
-import { mkdir, open, readFile, rename, rm, writeFile } from "node:fs/promises";
-import path from "node:path";
-
-export type VerificationEmailJob = {
-  jobId: string;
-  createdAt: string;
-  nextAttemptAt: string;
-  attempts: number;
-  maxAttempts: number;
-  userId: string;
-  email: string;
-  fullName: string;
-  verificationTokenId: string;
-  encryptedVerificationToken: string;
-};
-
-const defaultQueueFilePath = () =>
-  path.resolve(process.cwd(), ".tmp", "queues", "verification-email-jobs.json");
-const queueLockAcquireTimeoutMs = 5_000;
-const queueLockPollIntervalMs = 25;
-
-const isVerificationEmailJob = (value: unknown): value is VerificationEmailJob => {
-  if (!value || typeof value !== "object") {
-    return false;
-  }
-
-  const candidate = value as Partial<VerificationEmailJob>;
-  return (
-    typeof candidate.jobId === "string" &&
-    typeof candidate.createdAt === "string" &&
-    typeof candidate.nextAttemptAt === "string" &&
-    typeof candidate.attempts === "number" &&
-    typeof candidate.maxAttempts === "number" &&
-    typeof candidate.userId === "string" &&
-    typeof candidate.email === "string" &&
-    typeof candidate.fullName === "string" &&
-    typeof candidate.verificationTokenId === "string" &&
-    typeof candidate.encryptedVerificationToken === "string"
-  );
-};
-
-const readJobs = async ({ queueFilePath }: { queueFilePath: string }) => {
-  try {
-    const raw = await readFile(queueFilePath, "utf8");
-    const parsed = JSON.parse(raw) as unknown;
-
-    if (!Array.isArray(parsed)) {
-      return [] as VerificationEmailJob[];
-    }
-
-    return parsed.filter(isVerificationEmailJob);
-  } catch (error) {
-    const code = (error as NodeJS.ErrnoException).code;
-
-    if (code === "ENOENT") {
-      return [] as VerificationEmailJob[];
-    }
-
-    throw error;
-  }
-};
-
-const writeJobs = async ({
-  queueFilePath,
-  jobs,
-}: {
-  queueFilePath: string;
-  jobs: VerificationEmailJob[];
-}) => {
-  await mkdir(path.dirname(queueFilePath), { recursive: true });
-  const temporaryQueueFilePath = `${queueFilePath}.${randomBytes(8).toString("hex")}.tmp`;
-  let renamed = false;
-
-  try {
-    await writeFile(temporaryQueueFilePath, `${JSON.stringify(jobs, null, 2)}\n`, "utf8");
-    await rename(temporaryQueueFilePath, queueFilePath);
-    renamed = true;
-  } finally {
-    if (!renamed) {
-      await rm(temporaryQueueFilePath, { force: true }).catch(() => undefined);
-    }
-  }
-};
-
-const wait = (ms: number) => new Promise<void>((resolve) => setTimeout(resolve, ms));
-
-const acquireQueueLock = async (lockFilePath: string) => {
-  const startedAt = Date.now();
-
-  while (true) {
-    try {
-      const lockHandle = await open(lockFilePath, "wx");
-      await lockHandle.close();
-      return;
-    } catch (error) {
-      const code = (error as NodeJS.ErrnoException).code;
-
-      if (code !== "EEXIST") {
-        throw error;
-      }
-
-      if (Date.now() - startedAt >= queueLockAcquireTimeoutMs) {
-        throw new Error(`Timed out acquiring verification queue lock: ${lockFilePath}`);
-      }
-
-      await wait(queueLockPollIntervalMs);
-    }
-  }
-};
-
-const releaseQueueLock = async (lockFilePath: string) => {
-  await rm(lockFilePath, { force: true }).catch((error) => {
-    const code = (error as NodeJS.ErrnoException).code;
-    if (code !== "ENOENT") {
-      throw error;
-    }
-  });
-};
-
-const withQueueLock = async <T>({
-  queueFilePath,
-  operation,
-}: {
-  queueFilePath: string;
-  operation: () => Promise<T>;
-}) => {
-  const queueDirectory = path.dirname(queueFilePath);
-  const lockFilePath = `${queueFilePath}.lock`;
-
-  await mkdir(queueDirectory, { recursive: true });
-  await acquireQueueLock(lockFilePath);
-
-  try {
-    return await operation();
-  } finally {
-    await releaseQueueLock(lockFilePath);
-  }
-};
+import {
+  createVerificationEmailJob,
+  enqueueVerificationEmailJob as enqueueSharedVerificationEmailJob,
+  toBase64Url,
+  type VerificationEmailJob,
+} from "../../../../packages/shared/src/queues/verification-email.js";
 
 const deriveEncryptionKey = (secret: string) => createHash("sha256").update(secret, "utf8").digest();
 
-const toBase64Url = (value: Buffer) =>
-  value
-    .toString("base64")
-    .replaceAll("+", "-")
-    .replaceAll("/", "_")
-    .replace(/=+$/g, "");
+export { createVerificationEmailJob, type VerificationEmailJob };
 
 export const encryptVerificationToken = ({ token, secret }: { token: string; secret: string }) => {
   const iv = randomBytes(12);
@@ -156,45 +19,14 @@ export const encryptVerificationToken = ({ token, secret }: { token: string; sec
   return `${toBase64Url(iv)}.${toBase64Url(tag)}.${toBase64Url(encrypted)}`;
 };
 
-export const createVerificationEmailJob = ({
-  userId,
-  email,
-  fullName,
-  verificationTokenId,
-  encryptedVerificationToken,
-  now = new Date(),
-}: {
-  userId: string;
-  email: string;
-  fullName: string;
-  verificationTokenId: string;
-  encryptedVerificationToken: string;
-  now?: Date;
-}): VerificationEmailJob => ({
-  jobId: randomBytes(16).toString("hex"),
-  createdAt: now.toISOString(),
-  nextAttemptAt: now.toISOString(),
-  attempts: 0,
-  maxAttempts: 3,
-  userId,
-  email,
-  fullName,
-  verificationTokenId,
-  encryptedVerificationToken,
-});
-
 export const enqueueVerificationEmailJob = async ({
   job,
-  queueFilePath = defaultQueueFilePath(),
+  queueFilePath,
 }: {
   job: VerificationEmailJob;
   queueFilePath?: string;
 }) =>
-  withQueueLock({
+  enqueueSharedVerificationEmailJob({
+    job,
     queueFilePath,
-    operation: async () => {
-      const jobs = await readJobs({ queueFilePath });
-      jobs.push(job);
-      await writeJobs({ queueFilePath, jobs });
-    },
   });
